adds tests for coare_36 and restructures testing library, drops 3.8 support, and adds tests for 3.12 and 3.13
