+++ tests/data/c35_test_expected_old.csv
@@ -1,0 +1,16 @@
+tau,ustar,tstar,qstar,sensible,latent,buoyancy,webb,cd
+0.010839700299122611,0.0929537802403513,0.0035289594577549185,-5.8191668973998876e-05,-0.4150173752971719,16.955400714681762,0.7461770343390143,0.08888038490726081,0.0009527954787808854
+0.018315073623469603,0.12181259806571534,0.0043481156786216914,-6.563891859896353e-05,-0.6588128592234483,24.523223922626137,1.0590327687052603,0.19074225048503993,0.0009217693595408025
+0.028315449887385553,0.15272945389593492,0.005068413943322712,-6.985101255268265e-05,-0.9464477080296773,32.00884607308208,1.3463238377800446,0.3619101314251821,0.0009283349444582006
+0.21787092715226272,0.4270989073371195,0.004690421987323391,-4.934085866092562e-05,-2.4066331275054127,61.82797253900009,2.1206663593320934,0.9561985438017788,0.0022468002608987646
+0.2234095834055547,0.43635723213796795,0.0049312950732702514,-4.1506517863884125e-05,-2.5384631456603937,51.92827965063115,1.3475600403217147,1.073205627920829,0.001901212762637836
+0.0018574304086998044,0.03942350399437484,-0.016928100519770683,-8.055570715110489e-05,0.8443373584433258,9.95478134123862,1.5260927847593508,0.08893543701631644,0.0013997691524369152
+0.03763814634463486,0.1745265562167727,0.0052292787432463685,-5.820440100628473e-05,-1.1351992176602295,31.155970472302215,1.0472686270483724,0.22749589194731457,0.0012136759075838768
+0.054164075018491235,0.21115691044683293,0.0053829981664504385,-6.104002417071663e-05,-1.3897327076008326,38.67181086464837,1.3803024062224247,0.4195036265224866,0.0012341073406133135
+0.07897245326952188,0.25718634155848036,0.005292719833433635,-5.7082434516050114e-05,-1.6352957925193576,43.07251488379331,1.5186516284023626,0.6703771470595152,0.001345758773188982
+0.2630661209856086,0.47347804763047374,0.0050574660343511366,-4.173333740136162e-05,-2.824883337741888,56.65371858250518,1.4147653751544709,1.1628654086122279,0.0018503601023419824
+0.009672421531772488,0.08806757006817484,0.0002196987401259344,-4.9835878469862954e-05,-0.0244791387603142,13.757461947462694,0.9177037192563257,0.08266489425681092,0.0019016043928476256
+0.020426476922271772,0.1288236138067376,0.0018798653084120044,-5.417298418620115e-05,-0.30122497918760943,21.404356613871855,1.198144783278318,0.18010576541739998,0.0018224801002963596
+0.16138683513237076,0.3644660774764541,0.0041939541221669195,-4.901839564251868e-05,-1.8688848184347708,53.60320021866439,1.9706754064197005,0.5856135783776176,0.0027019000341239907
+0.19804467475656345,0.40726007945108816,0.004349379621473561,-4.7949277065028334e-05,-2.127985527961758,57.29328807583576,2.067265635115754,0.8965577349985705,0.0025841071817240035
+0.23873609027729353,0.4511299473878712,0.004363336717680007,-3.928605322152922e-05,-2.322138295222561,50.814249401305986,1.4805170725323866,1.0735505074796374,0.0025076059987283322
